move constant file
