+++ src/turbomind/utils/constant.h
